+++ app/routes/api_v1/endpoints/books.py
@@ -9,7 +9,9 @@ from starlette.status import HTTP_201_CREATED
 from app.controllers.books_controller import (
     get_books, get_book, create_book, update_book, delete_book)
 from app.schemas.books_schema import BooksBase, BooksAction
+from app.schemas.users_schema import UsersBase
 from app.settings.mysql_settings import SessionLocal
+from app.utils.auth import get_current_active_user
 
 #pylint: disable=invalid-name
 router = APIRouter()
@@ -28,8 +30,14 @@ def db_session() -> Generator:
 
 
 @router.get("/", response_model=List[BooksBase])
-def get_all_books(sql: Session = Depends(db_session)):
+def get_all_books(
+        sql: Session = Depends(db_session),
+        current_user: UsersBase = Depends(get_current_active_user)
+):
     """return books record"""
+    if current_user.Status:
+        raise HTTPException(status_code=400, detail="Inactive user")
+
     result = get_books(sql)
     return result
 
@@ -37,9 +45,13 @@ def get_all_books(sql: Session = Depends(db_session)):
 @router.get("/{book_id}", response_model=BooksBase)
 def get_book_by_id(
         book_id: int = Path(..., title="The Id of the book to get", ge=0),
-        sql: Session = Depends(db_session)
+        sql: Session = Depends(db_session),
+        current_user: UsersBase = Depends(get_current_active_user)
 ):
     """return a specific book record"""
+    if current_user.Status:
+        raise HTTPException(status_code=400, detail="Inactive user")
+
     result = get_book(sql, book_id=book_id)
 
     if result is None:
@@ -49,13 +61,20 @@ def get_book_by_id(
 
 
 @router.post("/", response_model=BooksAction, status_code=HTTP_201_CREATED)
-def add_new_book(newbook: BooksAction, sql: Session = Depends(db_session)):
+def add_new_book(
+        newbook: BooksAction,
+        sql: Session = Depends(db_session),
+        current_user: UsersBase = Depends(get_current_active_user)
+):
     """
     Create a book with all the information:
 
     - **Title**: must have a title
     - **AuthorId**: set the Id of the author (optional field)
     """
+    if current_user.Status:
+        raise HTTPException(status_code=400, detail="Inactive user")
+
     result = create_book(sql, book=newbook)
     return result
 
@@ -65,7 +84,8 @@ def update_book_by_id(
         book: BooksAction,
         book_id: int = Path(...,
                             title="The Id of the book to be updated", ge=0),
-        sql: Session = Depends(db_session)
+        sql: Session = Depends(db_session),
+        current_user: UsersBase = Depends(get_current_active_user)
 ):
     """
     update a book with all the information:
@@ -74,6 +94,9 @@ def update_book_by_id(
     - **Title**: must have a title
     - **AuthorId**: set the Id of the author if you want to update it (optional field)
     """
+    if current_user.Status:
+        raise HTTPException(status_code=400, detail="Inactive user")
+
     result = update_book(sql, book_id=book_id, book=book)
 
     if result is None:
@@ -86,9 +109,12 @@ def update_book_by_id(
 def delete_book_by_id(
         book_id: int = Path(...,
                             title="The Id of the book to be deleted", ge=0),
-        sql: Session = Depends(db_session)
+        sql: Session = Depends(db_session),
+        current_user: UsersBase = Depends(get_current_active_user)
 ):
     """delete a specific book"""
-    result = delete_book(sql, book_id=book_id)
+    if current_user.Status:
+        raise HTTPException(status_code=400, detail="Inactive user")
 
+    result = delete_book(sql, book_id=book_id)
     return result
