+++ app/routes/api_v1/endpoints/cars.py
@@ -8,7 +8,9 @@ from starlette.status import HTTP_201_CREATED
 
 from app.controllers.cars_controller import get_cars, get_car, create_car, update_car, delete_car
 from app.schemas.cars_schema import CarsBase, CarActions
+from app.schemas.users_schema import UsersBase
 from app.settings.mysql_settings import SessionLocal
+from app.utils.auth import get_current_active_user
 
 #pylint: disable=invalid-name
 router = APIRouter()
@@ -27,8 +29,14 @@ def db_session() -> Generator:
 
 
 @router.get("/", response_model=List[CarsBase])
-def get_all_cars(sql: Session = Depends(db_session)):
+def get_all_cars(
+        sql: Session = Depends(db_session),
+        current_user: UsersBase = Depends(get_current_active_user)
+):
     """return cars record"""
+    if current_user.Status:
+        raise HTTPException(status_code=400, detail="Inactive user")
+
     result = get_cars(sql)
     return result
 
@@ -36,9 +44,13 @@ def get_all_cars(sql: Session = Depends(db_session)):
 @router.get("/{car_id}", response_model=CarsBase)
 def get_car_by_id(
         car_id: int = Path(..., title="The Id of the car to get", ge=0),
-        sql: Session = Depends(db_session)
+        sql: Session = Depends(db_session),
+        current_user: UsersBase = Depends(get_current_active_user)
 ):
     """return a specific car record"""
+    if current_user.Status:
+        raise HTTPException(status_code=400, detail="Inactive user")
+
     result = get_car(sql, car_id=car_id)
 
     if result is None:
@@ -48,13 +60,20 @@ def get_car_by_id(
 
 
 @router.post("/", response_model=CarActions, status_code=HTTP_201_CREATED)
-def add_new_car(newcar: CarActions, sql: Session = Depends(db_session)):
+def add_new_car(
+        newcar: CarActions,
+        sql: Session = Depends(db_session),
+        current_user: UsersBase = Depends(get_current_active_user)
+):
     """
     Create a car with all the information:
 
     - **name**: must have a name
     - **price**: required
     """
+    if current_user.Status:
+        raise HTTPException(status_code=400, detail="Inactive user")
+
     result = create_car(sql, car=newcar)
     return result
 
@@ -63,7 +82,8 @@ def add_new_car(newcar: CarActions, sql: Session = Depends(db_session)):
 def update_car_by_id(
         car: CarActions,
         car_id: int = Path(..., title="The Id of the car to be updated", ge=0),
-        sql: Session = Depends(db_session)
+        sql: Session = Depends(db_session),
+        current_user: UsersBase = Depends(get_current_active_user)
 ):
     """
     update a car with all the information:
@@ -72,6 +92,9 @@ def update_car_by_id(
     - **name**: must have a name
     - **price**: required
     """
+    if current_user.Status:
+        raise HTTPException(status_code=400, detail="Inactive user")
+
     result = update_car(sql, car_id=car_id, car=car)
 
     if result is None:
@@ -83,9 +106,12 @@ def update_car_by_id(
 @router.delete("/{car_id}")
 def delete_car_by_id(
         car_id: int = Path(..., title="The Id of the car to be deleted", ge=0),
-        sql: Session = Depends(db_session)
+        sql: Session = Depends(db_session),
+        current_user: UsersBase = Depends(get_current_active_user)
 ):
     """delete a specific car"""
-    result = delete_car(sql, car_id=car_id)
+    if current_user.Status:
+        raise HTTPException(status_code=400, detail="Inactive user")
 
+    result = delete_car(sql, car_id=car_id)
     return result
